docs/schemas/requirements-governance.txt
pyyaml==6.0.3
jsonschema==4.25.1
yamllint==1.37.1
